Add idempotence
